Create TransactionInfo component
