+++ src/components/screens/reclaimBalance/modal/summary/index.js
@@ -5,9 +5,8 @@ import { selectAccount } from '@store';
 import { create } from '@api/transaction';
 import { tokenMap, MODULE_ASSETS_NAME_ID_MAP } from '@constants';
 import { useTransactionFeeCalculation, useTransactionPriority } from '@shared/transactionPriority';
-import AccountMigration from '@shared/accountMigration';
 import TransactionSummary from '@shared/transactionSummary';
-import LiskAmount from '@shared/liskAmount';
+import TransactionInfo from '@shared/transactionInfo';
 import { toRawLsk } from '@utils/lsk';
 import styles from './summary.css';
 
@@ -74,16 +73,7 @@ const Summary = ({
       token={tokenMap.LSK.key}
       classNames={styles.summaryContainer}
     >
-      <section>
-        <AccountMigration account={account.info.LSK} showBalance={false} />
-      </section>
-      <section>
-        <label>{t('Balance to reclaim')}</label>
-        <LiskAmount
-          val={Number(account.info.LSK.legacy.balance)}
-          token={tokenMap.LSK.key}
-        />
-      </section>
+      <TransactionInfo account={account} />
     </TransactionSummary>
   );
 };
